refactor: make controlled fields pass generic type to control for proper type inference for form values.
